Added clientes and pagos models and cliente route

+++ app/Repositories/PagoTarjetaRepository.php
@@ -4,6 +4,7 @@ namespace App\Repositories;
 
 use App\Models\PagoTarjeta;
 use App\Repositories\Contracts\RepositoryInterface;
+use Illuminate\Database\Eloquent\Collection;
 
 class PagoTarjetaRepository implements RepositoryInterface
 {
@@ -13,30 +14,36 @@ class PagoTarjetaRepository implements RepositoryInterface
         return PagoTarjeta::create($data);
     }
 
-    public function update($id, array $data): PagoTarjeta
+    public function update($id, array $data): ?PagoTarjeta
     {
-        $pagoTarjeta = PagoTarjeta::findOrFail($id);
-        $pagoTarjeta->update($data);
-        return $pagoTarjeta;
+        $pagoTarjeta = $this->findById($id);
+        if ($pagoTarjeta) {
+            $pagoTarjeta->update($data);
+            return $pagoTarjeta;
+        }
+        return null;
     }
 
-    public function delete($id): void
+    public function delete($id): bool
     {
-        $pagoTarjeta = PagoTarjeta::findOrFail($id);
-        $pagoTarjeta->delete();
+        $pagoTarjeta = $this->findById($id);
+        if ($pagoTarjeta) {
+            return $pagoTarjeta->delete();
+        }
+        return false;
     }
 
-    public function findById($id): PagoTarjeta
+    public function findById($id): ?PagoTarjeta
     {
-        return PagoTarjeta::findOrFail($id);
+        return PagoTarjeta::find($id);
     }
 
-    public function findByPaymentId(int $paymentId): PagoTarjeta
+    public function findByPaymentId(int $paymentId): ?PagoTarjeta
     {
-        return PagoTarjeta::where('id_pago', $paymentId)->firstOrFail();
+        return PagoTarjeta::where('id_pago', $paymentId)->first();
     }
 
-    public function getAll()
+    public function getAll(): Collection
     {
         return PagoTarjeta::all();
     }
